[tests] pytest plugins can now only be listet in top-level conftest
